Simplify (?) application init
