+++ renderer/src/orchestration.rs
@@ -1,1 +1,2 @@
+pub mod orchestrator;
 pub mod rendering;
